// estyle-server/src/modules/style/models/quiz/measurements/measurements.enum.ts
export enum VerticalBodyShape {
  LONGLEGSSHORTTORSO = 'longLegsShortTorso',
  SHORTLEGSLONGTORSO = 'shortLegsLongTorso',
  BALANCED = 'balanced',
}

export enum HorizontalBodyShape {
  HOURSGLASS = 'hourglass',
  TRIANGULAR = 'triangular',
  RECTANGULAR = 'rectangular',
  INVERTEDTRIANGULAR = 'invertedTriangular',
  DIAMOND = 'diamond',
  OVAL = 'oval',
}

export enum HeightCategory {
  PETITE = 'petite',
  SMALL = 'small',
  MEDIUM = 'medium',
  TALL = 'tall',
}

export enum NeckSize {
  SMALL = 'small',
  MEDIUM = 'medium',
  LONG = 'long',
}

export enum BMICategory {
  VERYTHIN = 'veryThin',
  NORMAL = 'normal',
  OVERWEIGHT = 'Overweight',
}